pants: cleanup stevedore_extensions plugin
move StevedoreNamespacesField to target_types
